Running into some bullshit bugs which dont make any fucking sense because my code is IDENTICAL to tims so I don't know what this fucking thing wants

+++ TimCoRetailManager/TRMData/dbo/Store Procedures/spSaleDetail_Insert.sql
@@ -2,7 +2,7 @@
 	@SaleId int,
 	@ProductId int,
 	@Quantity int,
-	@PurchasePrice int,
+	@PurchasePrice money,
 	@Tax money
 AS
 	begin
